Fix : Change several useRef to useState and something depending on them and others
in useEffect deps array, useEffect didn't work again as ref.current value was changed.
So I changed useRef to useState and other code lines.
But squareSize must be created by useRef because other setStates read the squareSize value asynchronously.
If I create squareSize by useState, it is read as previous value that is not changed yet in the moment expected as changed value.

And this fixing was done in resizing square that is my next commit