Fix : Create local var "isResizing" to decide to move square or not
in previous code I got and read the "selectedCornerForResizing" but in the moment it was the previous value and I needed to read the new value
or to know the upper code lines' result that decide to move square or not.
for that I created and used local variable "isResizing" and it works as I expected.

I also changed how to create the variable "selectedCornerForResizing" by from useState to useRef.
but now I think there was no need to change but I leave the code as it is cause there is also no need to change again.