Fix : Change the response data type and Get the image link properly
I don't need to set response type as detailed as one in dev tool's network tab cause I use both "transformResponse" in api setting and ".unwrap( )" in using api. I just set one property "link" as I need to get in response data type and that's all!

in previous code, I tried to read the data from "result.data.data.link". That was not an error in run time but was an error in actual data fetching. I wondered how I set and got the same data both in code editor and actual fetching, and now I got it. I could set and get data as so simple...

Plus this "result" is not the one I want to get. It just like state explanations after fetching.
[ const [ ... , result ] = useImageHostingMutation( ); ]

And now I also found that why I got the 400 error. It's because of "setSelectedProfileImage". It looks like after hosting image this setState changed the state and component rerendered and api was fired with wrong data. so I got the 400 error. now I get this again and I will fix it in next commit.