Fix : Initialize the values as trying to editing image not just once
I set selectedProfileImage as empty string to initialize the values on canvas modal for when trying to editing image not just once.
when I didn't, the values on the canvas modal were set as them as when the previous image was displayed and I could see the square did not fit in canvas and the feature for resizing square couldn't work.

(TL;DR)
as debugging, I guessed that the main reason that the variables in canvas modal had previous values might be that "REACT" library had previous value because of virtual dom system, even though in the moment I couldn't understand as my knoweldge about the virtual dom. anyway I started to looking for what variable is the problem : starting value x, y of square, square size, canvas size, image object... and I finally found out that the state "selectedProfileImage" has the previous value. I tried to solve this, and I did it! there were trials and my joy - sad loop... lol
(TL;DR end)

and I changed the state name from "isEditedImage" to "isEditingImage". previously when I clicked the cancel button on canvas modal the state "isEditedImage" was set as "true" but it didn't make sense because it is the time that I canceled editing image. so I changed the state name and other code lines depending on it.