Fix : Set "isLike" gotten from server
there have to be "toggleLikeResult.data" in deps array, useEffect.
I previously didn't because i was not sure if that was always considered as different when "toggleLikeResult.data" was same as changing multiple "isLike"s of multiple novelLists. and I'm still not sure of it. but without this code I couldn't set the "isLike" value so I did.
If there is a problem I will handle that later.