Fix : Create a variable to set all the novel list in DB and handle it
I fixed what I wrote in previous commit message description as other way not as I wrote.
since I set and handle a variable for all novel lists' titles and their simple info, the titles shown to user would be always the same whenever clicking other novel title. and also there would not need to delete a novel list where user canceled LIKE not to show it in the page.

and also to reuse data from server that is not necessary to get again, I did like this. I didn't change how to get and set and handle data from server.