Fix : Don't read data from server when there's no data
error had occurred as trying to read "userImg.src" when there was no data from server at first loading. so I handled it as displaying spinner.
but spinner is still required when there's some data in the page during fetching data. so I leave the exisiting code lines for it.