Fix : Display next list after canceling LIKE in login user's other's list
when I tried to get a next novel list right after canceling LIKE in login user's other's list, a request was done for current list not for next one. it still did when I removed tag id and set "if condition" in api slice.
it was because the tag "ContentsUpdated" used in mutation of toggleLike was the same as one in query for requesting list though tag id was different. so I divided the tag names to "ContentsUpdatedInHome" and "ContentsUpdatedInNovelList" and didn't set invalidatesTags of "ContentsUpdatedInNovelList" in other's list page of login user where the current list must not be refetched after canceling LIKE.
and as navigating to the next list page right after requesting I could display other novel list.