Fix : Use new query for getting all list titles of an user
it is required since I changed my code entirely in this component.
as I used cache data not kept it in local state, novel lists titles could be different whenever user clicked other novel list title. it's because other title list was set as "otherList" in current novel list data. so that would be different up to the moment when cache data came in. on the other hand, before I changed my code entirely, user could see the updated title list relatively since that was set as lastest combination of current list and "otherList" in it when it came from server at first.
anyway in previous code the title list of an user could be different for those reason, so I created and used new api for it. just for getting all novel list titles of user. and this data should be updated when login user canceled LIKE in his/her other's list page because in this case this novel list should be excluded.