Fix : Display the more button only when it is needed (complete)
when fixing bug I tried to handle the scrollable width of title list container as I configured before.  but there was no property of getBoundingClientRect() for it such as scrollWidth and I couldn't find proper way. finally I solved problem just as undoing some code and using state changed as argument of useComponentScrollWidth and
useComponentHeight.
so simple but spent quite time haha