Fix : Sort writings as using client states
before this change, I made states in parent component and passed it and its setState to the descendants. but now I change and use the state of sortType from global client states. and even more I use this for both components for mobile and tablet size devices. so if I change the browser size and the other component comes, the previous sort type remains and is used directly.