Fix : Display comment inputs depending on several things
many things should be different to display things properly such as whether the device type is mobile or not,  whether input is for root comment or reComment.
but in this changed code, the comment input in the FreeTalkDetail file is given the parent-comment-for-reComment in order to treat both root comment and reComment when device is mobile.
but some unexpected things happened in the related things that looked not good and perfect.
so I will change how to treat the comment input component by dividing it into several things depending on their use with certain args