Feat & Fix : Add a novel with the url and get the novel id and title
when user gives novel url to server, server collects novel info by going to the novel page with novel url.
and adds the info into DB ( if the info already exists in DB, updates in DB )
and returns the novel id and novel title to user.

I also adjusted things in add-writing page to work well