Fix : Maintain the scroll position in accumulated list when back
when going back to the list page, the scroll position and the accumulated list have to maintain with infinite scroll on mobile.
so I get the list updated only when adding or deleting a writing post and automatically going back to the list page.
though some changed data would not be updated, I couldn't help it.
 (I also checked that some popular website had the same problem)