Fix : Treat when going back and changing the search filter
in this case, the new list was not set nor displayed with infinite scroll on mobile in the previous code.
so I added a state to know whether the new list is in the progress for setting.
and also added code in the if-state for treating for when going back.
after changing the related code, I can handle better with the case when going back to the list page and searching for again