Fix : Get writing number with other api
so that user can get the writing posts and total writing number with all type at once whenever adding or deleting a writing in novel detail page.
this is a preparation for next commit to update data with provide and invalidate tags when there's a change with a writing post.