Refactor & Design : Remove "sample" from search type in novel
Instead, set the default search type to "Title" in novel search.
As the result, user can see the "Title" highlighted when enters the page first.
Also it is the same with the previous that the argument given to server is "sample" when the search word doesn't exist.
 "sample" is just for argument to server, not for component to show to users that makes users know what search type they can select.
So there's no change in function between server and client, but there's just a highlighted search type always so that user can catch this.