Fix : Reset search type if it is not connected with search category
When search params exist, "useResetFiltersFromUrlForSearchAll" checks if they are correct and change if they aren't.
(search params are added and used on desktop. when they don't exist on mobile or tablet, global states in filter slice work)
For this component to search for all, this addition is necessary.

And I removed the code for throwing an error where the search type is not right
 so that the "useResetFiltersFromUrlFor..." works and fixes the incorrect things in url
 and the page is displayed correctly