Fix : Fetch and accumulate the new list when back from detail page
In previous code, scrolling down automatically happens when going to the detail page as clicking one in the list.
As the result, the page number of search filter and isSettingTheList in filter slice changed at that moment,
 and when back to this page, the new list with next page number was fetched and replaced but not accumulated in the existing list.
So I fixed this as the next page number can only be set when the path is correct so that the problem can't happen when going to the detail page.

After back to this page, data was undefined and the scroll event couldn't happen in previous code.
So I made the scroll event would occur right after back to this page or when data existed.
When back here, user can scroll down and fetch the next list,
 and then isBackPageRef.current changes to false in order to fetch the list only when data would exist in the next time.

Also I let the previous list be as it is when user goes back here and fetches but data would be undefined.
It works too when user continues fetching data but data wouldn't exist in the latest request.