Design : Fit the bottom line in colored line
When a button for search type was selected, its bottom line highlighted with bright color.
But the end of bottom line overlap the border of parent component
 and then the part of straight bright line was put near the gray line of parent that was set with border-radius.
To make this look fittable, I gave the borer bottom radius set to the buttons which placed in the end between buttons.
Now the colored bottom line of selected button doesn't fight with the gray border line of parent component.