Fix : Add a case to display partner user img
Display partner user img when the message was the first unread message.
It's for the case where two messages are separated by the "unread message mark".
In the previous code, user img can't be displayed under the situation above where the create date and the user is the same between the previous and the current messages.
That didn't look good so I changed.