Fix : Change ChatRoomList and others after previous change
The previous change is joining all rooms right after login and getting new message in App component.
Now there is no need to get and join rooms and to treat new message in ChatRoomList component.

Change the unread message number of the current room to 0.
 It is the same to the new message that comes in the current room.
Get the rooms from chatSlice and then Sort and Display rooms.