Fix : Change code as getting messages with socket only
Use a new event to get messages with socket.
Treat messages identified roomId in chatSlice.
 Save and use messages through chatSlice and adjust them if needed.
Treat a new message coming to a room that the user is in or not considering the room exists or not.
Remove code that gets and sets data with an api.
Change to handle errors with socket connection.
etc