Fix : Scroll to a message only when entering the room
This change comes from the previous change to get messages with socket not an api.
And some additional change exists too.
Scrolling to a specific message works only when an user enters the room.
This doesn't work when a new message comes in here.
This will prevent the user from scrolling to the latest message automatically
 when he/she is reading messages.

It's okay to do so because I will add a feature to display a sign above message input that says a new message came in.
If the user clicks it, he/she will scroll automatically to the latest message.