Feat : Display a preview for new message to go there
This preview appears when a new message comes in that was sent by the partner.
If an user clicks it, he/she goes to the latest message with auto scrolling.
If the login user (not the partner) sends message, he/she goes to the latest message directly.
In addition, the preview doesn't appear when the y-scroll is near the bottom
  or disappears when the y-scroll is near bottom.

Also Change the message to read as one that was read last not the first unread one.