Fix : Change how to start to chat with an user
Replace an api to get chatroom id to go to chat with a socket event.
If an user clicks a message icon in an other user's page to chat,
 check that the room that they are joining exists or not.
If it does, go to the chatroom.
if not, send a socket event and get and set a new room
 and go to the chatroom.
And if the user send a new message in the room created just before,
 the partner user will receive a new message and set a new room and message in chatSlice.